[BUG] Move QuantileLoss regression test to test_metrics.py
